pyproject.toml
[project]
name = "fly-flower"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "ylei", email = "2457738122@qq.com"},
]
dependencies = [
    "fastapi>=0.110.1",
    "pydantic>=2.7.0",
    "uvicorn[standard]>=0.29.0",
    "tortoise-orm>=0.20.0",
    "aerich>=0.7.2",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.4.5",
    "httpx>=0.27.0",
    "pre-commit>=3.7.1",
]

[tool.pdm]
distribution = false


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.pdm.scripts]
# 开发环境
dev.env_file.override = ".env"
dev.cmd = "uvicorn flyflower.main:app --host 0.0.0.0 --port 8000 --reload"
post_install = "pre-commit install"

[tool.aerich]
tortoise_orm = "flyflower.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
